requirements.txt
annotated-types==0.7.0
black==24.10.0
blis==1.0.1
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
confection==0.1.5
cymem==2.0.8
flake8==7.1.1
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
langcodes==3.4.1
language_data==1.2.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
murmurhash==1.0.10
mypy-extensions==1.0.0
nltk==3.9.1
numpy==2.0.2
oauthlib==3.2.2
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
praw==7.7.1
prawcore==2.4.0
preshed==3.0.9
pycodestyle==2.12.1
pydantic==2.9.2
pydantic_core==2.23.4
pyflakes==3.2.0
Pygments==2.18.0
python-dotenv==1.0.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==1.3.1
rich==13.9.2
setuptools==75.2.0
shellingham==1.5.4
smart-open==7.0.5
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
textblob==0.18.0.post0
thinc==8.3.2
tqdm==4.66.5
tweepy==4.14.0
typer==0.12.5
typing_extensions==4.12.2
update-checker==0.18.0
urllib3==2.2.3
wasabi==1.1.3
weasel==0.4.1
websocket-client==1.8.0
wrapt==1.16.0
